% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stodom_functions.R
\name{ecdf_plot}
\alias{ecdf_plot}
\title{plot ecdfs}
\usage{
ecdf_plot(data_1, data_2, bins_size)
}
\arguments{
\item{data_1}{data 1.}

\item{data_2}{data 2.}

\item{bins_size}{bin size.}
}
\value{
The function returns a plot as a ggplot2 object.
}
\description{
This function computes the values of two empirical cumulative distribution function and plots the values.
}
\details{
This function computes the values of two empirical cumulative distribution function and plots the values.
}
\examples{

# load stodom
require(stodom)

 data_a <- rnorm(500, 3, 2)
 data_b <- rnorm(500, 1, 2)

# plot ecdfs
ecdf_plot(data_1 = data_a, data_2 = data_b, bins_size = 0.1)
}
